+++ R/vcov.cirls.R
@@ -4,7 +4,6 @@
 #
 ################################################################################
 
-#' @export
 vcov.cirls <- function(object, nsim = 1000, ...)
 {
   aliased <- summary(object)$aliased